Small Typos
Great work so far! All is explained very clearly and in great detail. Just a quick reminder that the backslash ``\`` is a ``Windows`` thing, and that ``Linux`` uses the regular slash ``/`` when specifying paths. I also added a few grammar improvements for you to consider changing.

Again great work, keep it up! 👍